submissions: change file structure
* Separates common/core files from author specific ones to prepare for
literature submission
