+++ src/submissions/common/__tests__/cleanupFormData.test.js
