+++ src/submissions/common/cleanupFormData.js
