+++ src/submissions/common/components/CollapsableForm/CollapsableForm.scss
