+++ src/submissions/common/components/SuggesterField.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/prop-types */
 
 import React, { Component } from 'react';
-import Suggester from '../../common/components/Suggester';
+import Suggester from '../../../common/components/Suggester';
 
 import withFormItem from '../withFormItem';
