+++ src/submissions/common/schemas/emptyObjectOrShapeOf.js
@@ -1,0 +1,11 @@
+import { object, lazy } from 'yup';
+import { isEmptyObjectShallow } from '../../../common/utils';
+
+export default function emptyObjectOrShapeOf(shape) {
+  return lazy(value => {
+    if (isEmptyObjectShallow(value)) {
+      return object();
+    }
+    return object().shape(shape);
+  });
+}
